Try to deal with veat sniping us.
One problem with veat is that we trigger on the first trade so don't see the trades it does right after on other markets and so we trade on incorrect information. One hack is if we see multiple markets have different probabilities than we thought after betting to sleep for some time.